Updated QuotationModel AmountRequired from int to decimal

+++ QuoteCalculator.Domain/Models/QuotationModel.cs
@@ -3,7 +3,8 @@
     public class QuotationModel
     {
         public Guid Id { get; set; } = new Guid();
-        public int AmountRequired { get; set; }
+        //public int AmountRequired { get; set; }
+        public decimal AmountRequired { get; set; }
         public int Term { get; set; }
         public required BorrowerModel Borrower { get; set; }
